refactor: move server/session tests into test/server/session
