+++ test/server/session/activity-hook.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { activityHookEffects, shouldNotifyTaskFinished } from "./activity-hook.js";
+import { activityHookEffects, shouldNotifyTaskFinished } from "../../../server/session/activity-hook.js";
 
 describe("activityHookEffects", () => {
   it("UserPromptSubmit sets working regardless of active", () => {
